[S05P13C105-62][BE] feat : 피드 삭제기능 추가

+++ backend/src/main/java/com/web/curation/feed/FeedDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface FeedDao extends JpaRepository<Feed, Long>{
+public interface FeedDao extends JpaRepository<Feed, Integer>{
 	@Query("SELECT f FROM Feed f join fetch f.photos ORDER BY f.writeDate DESC")
 	List<Feed> findAllJoinFetch();
 }
